pyproject.toml
[tool.poetry]
name = "ml-fairness"
version = "0.1.0"
description = ""
authors = ["christopher-cirelli <ccirelli2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fairlearn = "^0.9.0"
python-decouple = "^3.8"
pre-commit = "^3.5.0"
pylint = "^3.0.2"
pytest = "^7.4.3"
great-expectations = "^0.17.23"
gitpython = "^3.1.40"
mlflow = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
